# public/swagger.yaml
openapi: 3.0.0
info:
  description: 'Access token should be sent along with every request to SITE_NAME API in authorization header: `Authorization: Bearer <Token>` <br> This token can either be acquired via `auth/login` endpoint or from account settings page on [SITE_NAME website](SITE_URL/account/settings).'
  version: '1.0.0'
  title: SITE_NAME API
security:
  - accessToken: []
...
paths:
  /link:
    post:
      tags:
        - Links
      summary: Shorten a link
      operationId: createLink
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/CrupdateLinkPayload'
                - type: object
                  properties:
                    long_url:
                      type: string
                      example: https://google.com
                      description: Url to shorten
      responses:
        '200':
          description: On success, the response body contains the created link object in JSON format
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  link:
                    $ref: '#/components/schemas/Link'
        '401':
          $ref: '#/components/schemas/401-Response'
        '403':
          $ref: '#/components/schemas/403-Response'
        '422':
          description: Invalid data specified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/422-Response'
    get:
      tags:
        - Links
      summary: List all of your links
      operationId: getAllLinks
      parameters:
        - name: perPage
          in: query
          description: How many links to return per page
          schema:
            type: 'integer'
            default: 20
        - name: page
          in: query
          description: Which page to return
          schema:
            type: 'integer'
            default: 1
      responses:
        '200':
          description: Response body contains a paginated list of links in JSON format.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  pagination:
                    allOf:
                      - $ref: '#/components/schemas/Pagination'
                      - type: object
                        properties:
                          data:
                            type: array
                            items:
                              $ref: '#/components/schemas/Link'
        '401':
          $ref: '#/components/schemas/401-Response'
        '403':
          $ref: '#/components/schemas/403-Response'
...
  /link/{id}:
    parameters:
      - name: id
        in: path
        description: ID of the link
        example: 1
    get:
      tags:
        - Links
      summary: Get specified link details
      operationId: getLink
      responses:
        '200':
          description: Response body contains a link object in JSON format.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  link:
                    $ref: '#/components/schemas/Link'
        '401':
          $ref: '#/components/schemas/401-Response'
        '403':
          $ref: '#/components/schemas/403-Response'
    put:
      tags:
        - Links
      summary: Update specified link details
      operationId: updateLink
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrupdateLinkPayload'
      responses:
        '200':
          description: On success, the response body contains the updated link object in JSON format
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  link:
                    $ref: '#/components/schemas/Link'
        '401':
          $ref: '#/components/schemas/401-Response'
        '403':
          $ref: '#/components/schemas/403-Response'
        '422':
          description: Invalid data specified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/422-Response'
...
  /biolink:
    post:
      tags:
        - Biolinks
      summary: Create a new biolink
      operationId: createBiolink
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrupdateBiolinkPayload'
      responses:
        '200':
          description: On success, the response body contains the created biolink object in JSON format
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  biolink:
                    $ref: '#/components/schemas/Biolink'
        '401':
          $ref: '#/components/schemas/401-Response'
        '403':
          $ref: '#/components/schemas/403-Response'
        '422':
          description: Invalid data specified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/422-Response'
    get:
      tags:
        - Biolinks
      summary: List all of your biolinks
      operationId: getAllBiolinks
      parameters:
        - name: perPage
          in: query
          description: How many biolinks to return per page
          schema:
            type: integer
            default: 20
        - name: page
          in: query
          description: Which page to return
          schema:
            type: integer
            default: 1
      responses:
        '200':
          description: Response body contains a paginated list of link biolinks in JSON format.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  pagination:
                    allOf:
                      - $ref: '#/components/schemas/Pagination'
                      - type: object
                        properties:
                          data:
                            type: array
                            items:
                              $ref: '#/components/schemas/Biolink'
        '401':
          $ref: '#/components/schemas/401-Response'
        '403':
          $ref: '#/components/schemas/403-Response'
...
  /link-group/{id}/links:
    parameters:
      - name: id
        in: path
        description: ID of the link group
        example: 1
    get:
      tags:
        - Link Groups
      summary: List all links that are attached to specified group.
      operationId: getAllGroupLinks
      parameters:
        - name: perPage
          in: query
          description: How many links to return per page
          schema:
            type: integer
            default: 20
        - name: page
          in: query
          description: Which page to return
          schema:
            type: integer
            default: 1
      responses:
        '200':
          description: Response body contains a paginated list of links that are contained in specified group in JSON format.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  linkGroup:
                    $ref: '#/components/schemas/LinkGroup'
                  pagination:
                    allOf:
                      - $ref: '#/components/schemas/Pagination'
                      - type: object
                        properties:
                          data:
                            type: array
                            items:
                              $ref: '#/components/schemas/Link'
        '401':
          $ref: '#/components/schemas/401-Response'
        '403':
          $ref: '#/components/schemas/403-Response'
...
components:
  schemas:
    Pagination:
      type: object
      properties:
        current_page:
          type: integer
          format: int64
          example: 1
        from:
          type: integer
          format: int64
          example: 1
        to:
          type: integer
          format: int64
          example: 100
        per_page:
          type: integer
          format: int64
          example: 25
        last_page:
          type: integer
          format: int64
          example: 156
        total:
          type: integer
          format: int64
          example: 264
    Link:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        title:
          type: string
          example: 'reddit: the front page of the internet'
          description: Title for link. Either manually entered or extracted from destination url meta tags.
        hash:
          type: string
          example: rF7r6
          description: Randomly generated short ID for the link
        alias:
          type: string
          example: red
          description: Manually entered short ID for the link (if was entered)
        long_url:
          type: string
          example: https://reddit.com
          description: Orignal url that was shortened
        short_url:
          type: string
          example: SITE_URL/rF7r6
          description: Full short url for link. Includes protocol, domain and hash/alias
        disabled:
          type: boolean
          example: false
          description: Whether link is enabled or not. Disabled links will show a not found page instead of redirecting to the destination url.
        expires_at:
          type: string
          example: '2021-05-24 18:03:59'
          description: Expiration date for the link. After this date not found page will be shown instead of redirecting to destination url.
        description:
          type: string
          example: "Reddit is a network of communities based on people's interests. Find communities you're interested in, and become part of an online community!"
          description: Description for link. Either manually entered or extracted from destination url meta tags.
        type:
          type: string
          example: direct
          description: Link redirection type. Direct, frame, overlay, splash, or page.
        created_at:
          type: string
          example: '2021-05-24 18:03:59'
          description: Date link was created at.
        updated_at:
          type: string
          example: '2021-05-24 18:03:59'
          description: Date link was last updated.
        clicked_at:
          type: string
          example: '2021-05-24 18:03:59'
          description: Date link was last clicked / visited.
        deleted_at:
          type: string
          example: '2021-05-24 18:03:59'
          description: Date link was moved to trash (if link is trashed currently)
        image:
          type: string
          example: https://www.redditstatic.com/icon.png
          description: Url to the image for the link. Either manually entered/uploaded or extracted from destination url meta tags.
        clicks_count:
          type: number
          example: 854
          description: Total number of clicks/visists for the link.
        expiration_clicks:
          type: number
          example: 10
          description: After how many clicks/visits will the link expire.
        utm:
          type: string
          example: 'utm_source=google&utm_medium=banner'
          description: UTM tags added to link via utm builder.
        has_password:
          type: boolean
          default: false
          description: Whether link is password protected
        rules:
          nullable: true
          type: array
          description: Rules attached to the link.
          items:
            $ref: '#/components/schemas/LinkRule'
        tags:
          nullable: true
          type: array
          description: Tags attached to the link.
          items:
            $ref: '#/components/schemas/Tag'
        pixels:
          nullable: true
          type: array
          description: Tracking pixels attached to the link.
          items:
            $ref: '#/components/schemas/TrackingPixel'
        groups:
          nullable: true
          type: array
          description: Groups this link is attached to.
          items:
            $ref: '#/components/schemas/LinkGroup'
    CrupdateLinkPayload:
      type: object
      properties:
        alias:
          type: string
          nullable: true
          example: short
          description: Custom short ID for the link
        type:
          type: string
          nullable: true
          default: direct
          description: Link redirection type. Direct, frame, overlay, splash, or page.
        password:
          type: string
          nullable: true
          default: null
          description: New password, if link should be password protected.
        active:
          type: boolean
          nullable: true
          default: true
          description: Whether link should be disabled or not.
        expires_at:
          type: string
          nullable: true
          example: '2022-05-06'
          description: 'Format: "Y-m-d H:i:s" or "Y-m-d". Expiration date for the link. After this date not found page will be shown instead of redirecting to destination url.'
        activates_at:
          type: string
          nullable: true
          example: '2022-05-06'
          description: 'Format: "Y-m-d H:i:s" or "Y-m-d". Activation date for the link. Link will only become accessible after this date.'
        utm:
          type: string
          nullable: true
          example: 'utm_source=google&utm_medium=banner'
          description: UTM query string that should be attached to link.
        domain_id:
          type: number
          nullable: true
          default: null
          description: ID of the domain link should be accessible at. Null for all domains, 0 for main site domain only.
        title:
          type: string
          nullable: true
          example: Google
          description: Ttitle for the link. If empty, it will be extracted from destination url meta tags.
        description:
          type: string
          nullable: true
          example: Search Engine
          description: Description for the link. If empty, it will be extracted from destination url meta tags.
        pixels:
          type: array
          nullable: true
          items:
            type: integer
            example: 495
          description: Array of pixel IDs that should be attached to this link.
        groups:
          type: array
          nullable: true
          items:
            type: integer
            example: 54
          description: Array of group IDs that this link should be attached to.
        rules:
          nullable: true
          type: array
          description: Array of link rules that should be created and attached to this link.
          items:
            $ref: '#components/schemas/LinkRule'
    LinkGroup:
      type: object
      properties:
        name:
          type: string
          example: Social Media
          description: Name for the group.
        active:
          type: boolean
          example: false
          description: Whether this group is publicly acessible via its url.
        hash:
          type: string
          example: my-group
          description: Randomly generated or manually entered slug for the group.
        rotator:
          type: boolean
          example: false
          description: Whether it's a rotator group. Rotator group will redirect to random url contained in the group.
        description:
          type: string
          example: My group containing social media links
          description: Description for the group.
        created_at:
          type: string
          example: '2021-05-24 18:03:59'
          description: Date group was created at.
        updated_at:
          type: string
          example: '2021-05-24 18:03:59'
          description: Date group was last updated.
...
    Biolink:
      type: object
      properties:
        name:
          type: string
          example: Social Media
          description: Name for the biolink.
        active:
          type: boolean
          example: false
          description: Whether this biolink is publicly acessible via its url.
        hash:
          type: string
          example: my-group
          description: Randomly generated or manually entered slug for the biolink.
        expires_at:
          type: string
          nullable: true
          example: '2022-05-07'
          description: 'Format: "Y-m-d H:i:s" or "Y-m-d". Expiration date for the biolink. After this date biolink will no longer be accessible.'
        activates_at:
          type: string
          nullable: true
          example: '2022-05-06'
          description: 'Format: "Y-m-d H:i:s" or "Y-m-d". Activation date for the biolink. Biolink will only become accessible after this date.'
        utm:
          type: string
          nullable: true
          example: 'utm_source=google&utm_medium=banner'
          description: UTM query string that should be attached to biolink url.
        password:
          type: string
          nullable: true
          default: null
          description: New password, if biolink should be password protected.
        domain_id:
          type: number
          nullable: true
          default: null
          description: ID of the domain biolink should be accessible at. Null for all domains, 0 for main site domain only.
        description:
          type: string
          example: My biolink containing social media links
          description: Description for the biolink.
        created_at:
          type: string
          example: '2021-05-24 18:03:59'
          description: Date biolink was created at.
        updated_at:
          type: string
          example: '2021-05-24 18:03:59'
          description: Date biolink was last updated.
    CrupdateBiolinkPayload:
      type: object
      properties:
        name:
          type: string
          example: Social Media
          description: Name for the biolink.
        active:
          type: boolean
          example: false
          description: Whether this biolink is publicly acessible via its url.
        hash:
          type: string
          example: my-biolink
          description: Randomly generated or manually entered slug for the biolink.
        expires_at:
          type: string
          nullable: true
          example: '2022-05-06'
          description: 'Format: "Y-m-d H:i:s" or "Y-m-d". Expiration date for the biolink. After this date biolink will no longer be accessible.'
        activates_at:
          type: string
          nullable: true
          example: '2022-05-06'
          description: 'Format: "Y-m-d H:i:s" or "Y-m-d". Activation date for the biolink. Biolink will only become accessible after this date.'
        utm:
          type: string
          nullable: true
          example: 'utm_source=google&utm_medium=banner'
          description: UTM query string that should be attached to biolink url.
        password:
          type: string
          nullable: true
          default: null
          description: New password, if biolink should be password protected.
        domain_id:
          type: number
          nullable: true
          default: null
          description: ID of the domain biolink should be accessible at. Null for all domains, 0 for main site domain only.
        description:
          type: string
          example: My biolink containing social media links
          description: Description for the biolink.
        pixels:
          type: array
          nullable: true
          items:
            type: integer
            example: 495
          description: Array of pixel IDs that should be attached to this biolink.
        rules:
          nullable: true
          type: array
          description: Array of link rules that should be created and attached to this biolink.
          items:
            $ref: '#components/schemas/LinkRule'
...
    LinkRule:
      type: object
      properties:
        type:
          type: string
          example: geo
          description: Rule type. Geo, platform or device.
        key:
          type: string
          example: us
          description: A condition for rule. For example country code for geo rules.
        value:
          type: string
          example: https://facebook.com
          description: Destination url use should be redirect to if the rule condition is matched.
...
    TrackingPixel:
      type: object
      properties:
        name:
          type: string
          example: Facebook Pixel
          description: Name of the tracking pixel.
        type:
          type: string
          example: facebook
          description: Type of pixel. Usually the name of tracking service (facebook, twitter, adwords etc.)
        pixel_id:
          type: string
          example: 4564654564
          description: Tracking service ID for this pixel
        head_code:
          type: string
          description: Custom code that should be injected into <head> tag, if provided by tracking service.
        body_code:
          type: string
          description: Custom code that should be injected into <body> tag, if provided by tracking service.
...
    Tag:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: favorites
    401-Response:
      description: Unauthenticated. Either access token is not provided or is invalid.
    403-Response:
      description: Unauthorized access. You don't have permissions required to perform this operation.
...
    422-Response:
      type: object
      properties:
        status:
          type: string
          example: error
        message:
          type: string
          example: 'Reason for the error'
        errors:
          type: object
          properties:
            some_data_1:
              type: string
              example: Error message for data 1
            some_data_2:
              type: string
              example: Error message for data 2
  securitySchemes:
    accessToken:
      type: http
      scheme: bearer